more CE to be fixed

+++ rustfs/src/constants.rs
@@ -1,8 +1,8 @@
 pub const BLOCK_SIZE: usize = 512;
 pub const LIST_SIZE: usize = 16;
 pub const INODE_SIZE: usize = 32;
-pub const DIR_TYPE: u64 = 2;
-pub const FILE_TYPE: u64 = 1;
+pub const DIR_TYPE: usize = 2;
+pub const FILE_TYPE: usize = 1;
 pub const O_RDONLY: u32 = (1 << 0);
 pub const O_WRONLY: u32 = (1 << 1);
 pub const O_RDWR: u32 = (1 << 2);
